# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cudium API Docs",
        "version": "1.0"
    },
    "paths": {
...
        "/api/pin": {
            "post": {
                "tags": [
                    "Pin"
                ],
                "summary": "Create User's Pin",
                "description": "allows user to be able to create pin",
                "operationId": "addpin",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9hdXRoXC9yZWdpc3RlciIsImlhdCI6MTYxOTAzMjk2MywiZXhwIjoxNjE5MDM2NTYzLCJuYmYiOjE2MTkwMzI5NjMsImp0aSI6IllsMmZ1dGRxaksybWNjcVgiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.ds1SkBXLeFU_cx0EVx8m5l3pc3IBA4_jexYv6TC7yzs"
                    },
                    {
                        "name": "pin",
                        "in": "query",
                        "description": "User's 4-digit Pin",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Error Occurred,check error object"
                    }
                }
            }
        },
        "/api/profile/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Let you view user's profile",
                "description": "Allow to view user's profile",
                "operationId": "viewUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id to get user",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYxOTE2MTA2OCwiZXhwIjoxNjE5MTY0NjY4LCJuYmYiOjE2MTkxNjEwNjgsImp0aSI6InYwaUFvTEdldnI2VlQ4bXIiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dJx2G6MdbDz4r9SSgDdKpIbBp3gNtuFt1AuvbwQmQoM"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Retrieve order"
                    },
                    "400": {
                        "description": "Error occurred"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/profile": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Edit user's profile",
                "description": "allows user to be able to edit profile",
                "operationId": "editProfile",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9hdXRoXC9yZWdpc3RlciIsImlhdCI6MTYxOTAzMjk2MywiZXhwIjoxNjE5MDM2NTYzLCJuYmYiOjE2MTkwMzI5NjMsImp0aSI6IllsMmZ1dGRxaksybWNjcVgiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.ds1SkBXLeFU_cx0EVx8m5l3pc3IBA4_jexYv6TC7yzs"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User's new phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's  new Password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Error Occurred"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}